refactor(crdt): establish canonical mutation contracts
