+++ packages/shared-server/rallar-system/services/crdt-compact-snapshot.ts
@@ -1,11 +1,9 @@
 import type { RallarCrdtSnapshotEnvelope } from '@shared/crdt/mod.ts';
 
-import type { CrdtCanonicalSnapshotEnvelope } from './crdt-mutation-contracts.ts';
+import type { CrdtCanonicalSnapshotEnvelope } from '../crdt/mutation/crdt-mutation-contracts.ts';
 import { requireString } from './exact-object-codec.ts';
 
-export function requireCrdtCanonicalSnapshotReason(
-  value: unknown,
-): asserts value is string {
+export function requireCrdtCanonicalSnapshotReason(value: unknown): asserts value is string {
   requireString(value, 'snapshot reason');
   if (value.trim().length === 0) {
     throw new TypeError('snapshot reason must contain a non-whitespace character');
